fixed issues where 0 minutes was cuasing problems with the total hours fraction
